rename Skeleton -> Template
